new options -bin -vbin -sigma etc.
Now saving png is optimized for display, not for archiving. The amean is set to 127 and 1.5 sigma is used as the upper cutoff. -sigma changes the upper limit. -bin does a simple binning of nxn pixels. -vbin does a "vertical" binning of adjacent frames. Any remainder of the frames will not be used in binning. Eg. 30 frames, -vbin 4 will only bin the first 28 frames in 7 resulting pngs.